backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
PyMuPDF>=1.24.8
python-docx==1.1.2
vaderSentiment==3.3.2
email-validator>=2.0.0
firebase-admin>=6.0.0
httpx>=0.24.0
deep-translator>=1.11.4

